Add draft.js css to properly show lists. Refactor toolbar and stylebuttons into separate files
